handle expires_in RFC6749 violation (#27)
Site24x7 specifies the value in the expires_in field of the token
response in milliseconds although it must be provided in seconds (e.g.
3600000 instead of 3600 for a validity of one hour). This is a violation
of RFC6749 (see https://tools.ietf.org/html/rfc6749#section-5.1) and
prevents the token refresh mechanism from functioning properly as an
actual validity of one hour is returned as 41.6 days by the Site24x7
OAuth Authorization Server. This makes an instance of the API client
unusable after running for longer than an hour.

This patch implements a token source which handles malformed token
expiry dates and fixes them. It also sets the `AuthStyle` for the OAuth
endpoint to `oauth2.AuthStyleInParams` to avoid the probing round trip.