Work around bug in Site24x7 API that returns severity as a string (#30)
The API still accepts integers as input, but returns the severity value
as a string. Unmarshalling fails. Just adding `"severity,string"` in
the struct is not ok because that would also marshal the field as
a string.

The custom unmarshaller will handle convert from a string to an
integer if needed. In case Site24x7 fixes their API, this will continue
working.

This fixes: https://github.com/Bonial-International-GmbH/terraform-provider-site24x7/issues/19